blobs/raspberrypi: Revert start_cd.elf to f4e3f0f revision
The current HEAD revision of start.elf breaks msd
mode for custom raspberrypi devices. All files are
transfered to the device but it does not enter mass
storage mode.

We have an open issue in upstream rpi usbboot,
so this patch should not be merged, we'll only
use it to unblock testing.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>